requirements.txt
Django==1.10.2
django-allauth==0.28.0
django-debug-toolbar==1.6
django-filter==0.15.3
djangorestframework==3.5.3
gunicorn==19.6.0
Markdown==2.6.7
oauthlib==2.0.0
Pillow==3.4.2
pkg-resources==0.0.0
python-openid==2.2.5
requests==2.11.1
requests-oauthlib==0.7.0
sqlparse==0.2.2
